requirements.txt
asgiref==3.5.0
dj-rest-auth==2.2.4
Django==4.0.3
django-admin==2.0.1
django-cors-headers==3.11.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-six==1.0.4
django-webpack-loader==1.4.1
djangorestframework==3.13.1
Markdown==3.3.6
pytz==2022.1
screen==1.0.1
sqlparse==0.4.2
xlwt==1.3.0
